disable pg wire protocol and cache in default
