+++ wren-server/src/main/java/io/wren/server/module/WebModule.java
@@ -17,9 +17,12 @@ package io.wren.server.module;
 import com.google.inject.Binder;
 import com.google.inject.Scopes;
 import io.airlift.configuration.AbstractConfigurationAwareModule;
+import io.wren.base.config.PostgresWireProtocolConfig;
 import io.wren.main.PreviewService;
 import io.wren.main.ValidationService;
+import io.wren.main.pgcatalog.NoOpPgCatalogManager;
 import io.wren.main.pgcatalog.PgCatalogManager;
+import io.wren.main.pgcatalog.PgCatalogManagerImpl;
 import io.wren.main.web.AnalysisResource;
 import io.wren.main.web.CacheResource;
 import io.wren.main.web.ConfigResource;
@@ -36,15 +39,21 @@ public class WebModule
     @Override
     protected void setup(Binder binder)
     {
+        PostgresWireProtocolConfig config = buildConfigObject(PostgresWireProtocolConfig.class);
         jaxrsBinder(binder).bind(MDLResource.class);
         jaxrsBinder(binder).bind(LineageResource.class);
-        jaxrsBinder(binder).bind(CacheResource.class);
         jaxrsBinder(binder).bind(AnalysisResource.class);
         jaxrsBinder(binder).bind(ConfigResource.class);
         jaxrsBinder(binder).bind(DuckDBResource.class);
         jaxrsBinder(binder).bindInstance(new WrenExceptionMapper());
         binder.bind(PreviewService.class).in(Scopes.SINGLETON);
-        binder.bind(PgCatalogManager.class).in(Scopes.SINGLETON);
         binder.bind(ValidationService.class).in(Scopes.SINGLETON);
+        if (config.isPgWireProtocolEnabled()) {
+            jaxrsBinder(binder).bind(CacheResource.class);
+            binder.bind(PgCatalogManager.class).to(PgCatalogManagerImpl.class).in(Scopes.SINGLETON);
+        }
+        else {
+            binder.bind(PgCatalogManager.class).to(NoOpPgCatalogManager.class).in(Scopes.SINGLETON);
+        }
     }
 }
